tarea3/backend/requirements.txt
azure-functions
azure-servicebus==7.14.1
pydantic==2.8.2
requests
